+++ include/rockyou/pch.h
@@ -13,10 +13,13 @@
 #include <cstdint>
 #include <cstdio>
 #include <cstring>
+#include <charconv>
 #include <filesystem>
+#include <fstream>
 #include <iomanip>
 #include <iostream>
 #include <format>
+#include <limits>
 #include <map>
 #include <mutex>
 #include <optional>